Minor refactoring to make detect happy: Split CompanyManager.kt into a read-only service CompanyQueryManager and an update-only service CompanyAlterationManager.
