+++ dataland-backend/src/main/kotlin/org/dataland/datalandbackend/controller/CompanyDataController.kt
@@ -11,7 +11,8 @@ import org.dataland.datalandbackend.model.StoredCompany
 import org.dataland.datalandbackend.model.enums.company.IdentifierType
 import org.dataland.datalandbackend.repositories.CompanyIdentifierRepository
 import org.dataland.datalandbackend.repositories.utils.StoredCompanySearchFilter
-import org.dataland.datalandbackend.services.CompanyManager
+import org.dataland.datalandbackend.services.CompanyAlterationManager
+import org.dataland.datalandbackend.services.CompanyQueryManager
 import org.dataland.datalandbackendutils.exceptions.ResourceNotFoundApiException
 import org.dataland.keycloakAdapter.auth.DatalandAuthentication
 import org.slf4j.LoggerFactory
@@ -27,15 +28,16 @@ import org.springframework.web.bind.annotation.RestController
 
 @RestController
 class CompanyDataController(
-    @Autowired var companyManager: CompanyManager,
-    @Autowired var companyIdentifierRepository: CompanyIdentifierRepository,
+    @Autowired private val companyAlterationManager: CompanyAlterationManager,
+    @Autowired private val companyQueryManager: CompanyQueryManager,
+    @Autowired private val companyIdentifierRepository: CompanyIdentifierRepository,
 ) : CompanyApi {
     private val logger = LoggerFactory.getLogger(javaClass)
 
     override fun postCompany(companyInformation: CompanyInformation): ResponseEntity<StoredCompany> {
         logger.info("Received a request to post a company with name '${companyInformation.companyName}'")
         return ResponseEntity.ok(
-            companyManager.addCompany(companyInformation)
+            companyAlterationManager.addCompany(companyInformation)
                 .toApiModel(DatalandAuthentication.fromContext()),
         )
     }
@@ -54,7 +56,7 @@ class CompanyDataController(
                 "onlyWithDataFromCurrentUser='$onlyWithDataFromCurrentUser'",
         )
         return ResponseEntity.ok(
-            companyManager.searchCompaniesAndGetApiModel(
+            companyQueryManager.searchCompaniesAndGetApiModel(
                 StoredCompanySearchFilter(
                     searchString = searchString ?: "",
                     nameOnlyFilter = onlyCompanyNames,
@@ -72,7 +74,7 @@ class CompanyDataController(
         searchString: String,
     ): ResponseEntity<List<CompanyIdAndName>> {
         return ResponseEntity.ok(
-            companyManager.searchCompaniesByNameOrIdentifierAndGetApiModel(
+            companyQueryManager.searchCompaniesByNameOrIdentifierAndGetApiModel(
                 searchString,
             ),
         )
@@ -93,15 +95,15 @@ class CompanyDataController(
     override fun getAvailableCompanySearchFilters(): ResponseEntity<CompanyAvailableDistinctValues> {
         return ResponseEntity.ok(
             CompanyAvailableDistinctValues(
-                countryCodes = companyManager.getDistinctCountryCodes(),
-                sectors = companyManager.getDistinctSectors(),
+                countryCodes = companyQueryManager.getDistinctCountryCodes(),
+                sectors = companyQueryManager.getDistinctSectors(),
             ),
         )
     }
 
     override fun getCompanyById(companyId: String): ResponseEntity<StoredCompany> {
         return ResponseEntity.ok(
-            companyManager
+            companyQueryManager
                 .getCompanyApiModelById(companyId, DatalandAuthentication.fromContextOrNull()),
         )
     }
@@ -110,9 +112,9 @@ class CompanyDataController(
         companyId: String,
         companyInformationPatch: CompanyInformationPatch,
     ): ResponseEntity<StoredCompany> {
-        companyManager.patchCompany(companyId, companyInformationPatch)
+        companyAlterationManager.patchCompany(companyId, companyInformationPatch)
         return ResponseEntity.ok(
-            companyManager
+            companyQueryManager
                 .getCompanyApiModelById(companyId, DatalandAuthentication.fromContextOrNull()),
         )
     }
@@ -121,14 +123,14 @@ class CompanyDataController(
         companyId: String,
         companyInformation: CompanyInformation,
     ): ResponseEntity<StoredCompany> {
-        companyManager.putCompany(companyId, companyInformation)
+        companyAlterationManager.putCompany(companyId, companyInformation)
         return ResponseEntity.ok(
-            companyManager
+            companyQueryManager
                 .getCompanyApiModelById(companyId, DatalandAuthentication.fromContextOrNull()),
         )
     }
 
     override fun getTeaserCompanies(): List<String> {
-        return companyManager.getTeaserCompanyIds()
+        return companyQueryManager.getTeaserCompanyIds()
     }
 }
